use passed down api information as props so that it can be re-used in the future
